# Introduce IPropertyDescriptor and specialized descriptors