--- a/Jint/Native/Argument/ArgumentsInstance.cs
+++ b/Jint/Native/Argument/ArgumentsInstance.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using Jint.Native.Function;
 using Jint.Native.Object;
 using Jint.Runtime;
@@ -14,6 +15,12 @@ namespace Jint.Native.Argument
     /// </summary>
     public class ArgumentsInstance : ObjectInstance
     {
+        // cache key container for array iteration for less allocations
+        private static readonly ThreadLocal<List<string>> _mappedNamed = new ThreadLocal<List<string>>(() => new List<string>());
+
+        private readonly Action<ArgumentsInstance> _initializer;
+        private bool _initialized;
+
         private ArgumentsInstance(Engine engine, Action<ArgumentsInstance> initializer) : base(engine)
         {
             _initializer = initializer;
@@ -22,12 +29,9 @@ namespace Jint.Native.Argument
 
         public bool Strict { get; set; }
 
-        private Action<ArgumentsInstance> _initializer;
-        private bool _initialized;
-
         protected override void EnsureInitialized()
         {
-            if(_initialized)
+            if (_initialized)
             {
                 return;
             }
@@ -39,37 +43,36 @@ namespace Jint.Native.Argument
 
         public static ArgumentsInstance CreateArgumentsObject(Engine engine, FunctionInstance func, string[] names, JsValue[] args, EnvironmentRecord env, bool strict)
         {
-            var obj = new ArgumentsInstance(engine, self =>
+            void Initializer(ArgumentsInstance self)
             {
                 var len = args.Length;
                 self.SetOwnProperty("length", new NonEnumerablePropertyDescriptor(len));
-                var map = engine.Object.Construct(Arguments.Empty);
-                var mappedNamed = new List<string>();
-                var indx = 0;
-                while (indx <= len - 1)
+                if (args.Length > 0)
                 {
-                    var indxStr = TypeConverter.ToString(indx);
-                    var val = args[indx];
-                    self.SetOwnProperty(indxStr, new ConfigurableEnumerableWritablePropertyDescriptor(val));
-                    if (indx < names.Length)
+                    var map = engine.Object.Construct(Arguments.Empty);
+                    var mappedNamed = _mappedNamed.Value;
+                    mappedNamed.Clear();
+                    for (var indx = 0; indx < len; indx++)
                     {
-                        var name = names[indx];
-                        if (!strict && !mappedNamed.Contains(name))
+                        var indxStr = TypeConverter.ToString(indx);
+                        var val = args[indx];
+                        self.SetOwnProperty(indxStr, new ConfigurableEnumerableWritablePropertyDescriptor(val));
+                        if (indx < names.Length)
                         {
-                            mappedNamed.Add(name);
-                            Func<JsValue, JsValue> g = n => env.GetBindingValue(name, false);
-                            var p = new Action<JsValue, JsValue>((n, o) => env.SetMutableBinding(name, o, true));
-
-                            map.SetOwnProperty(indxStr, new ClrAccessDescriptor(engine, g, p) { Configurable = true });
+                            var name = names[indx];
+                            if (!strict && !mappedNamed.Contains(name))
+                            {
+                                mappedNamed.Add(name);
+                                map.SetOwnProperty(indxStr, new ClrAccessDescriptor(env, engine, name));
+                            }
                         }
                     }
-                    indx++;
-                }
 
-                // step 12
-                if (mappedNamed.Count > 0)
-                {
-                    self.ParameterMap = map;
+                    // step 12
+                    if (mappedNamed.Count > 0)
+                    {
+                        self.ParameterMap = map;
+                    }
                 }
 
                 // step 13
@@ -84,7 +87,9 @@ namespace Jint.Native.Argument
                     self.DefineOwnProperty("caller", new PropertyDescriptor(get: thrower, set: thrower, enumerable: false, configurable: false), false);
                     self.DefineOwnProperty("callee", new PropertyDescriptor(get: thrower, set: thrower, enumerable: false, configurable: false), false);
                 }
-            });
+            }
+
+            var obj = new ArgumentsInstance(engine, Initializer);
 
             // These properties are pre-initialized as their don't trigger
             // the EnsureInitialized() event and are cheap
@@ -92,20 +97,12 @@ namespace Jint.Native.Argument
             obj.Extensible = true;
             obj.Strict = strict;
 
-
             return obj;
         }
 
         public ObjectInstance ParameterMap { get; set; }
 
-        public override string Class
-        {
-            get
-            {
-                return "Arguments";
-            }
-        }
-
+        public override string Class => "Arguments";
 
         public override IPropertyDescriptor GetOwnProperty(string propertyName)
         {
@@ -163,7 +160,7 @@ namespace Jint.Native.Argument
             if (desc.IsAccessorDescriptor())
             {
                 var setter = desc.Set.TryCast<ICallable>();
-                setter.Call(JsValue, new[] { value });
+                setter.Call(JsValue, new[] {value});
             }
             else
             {
@@ -188,6 +185,7 @@ namespace Jint.Native.Argument
                         throw new JavaScriptException(Engine.TypeError);
                     }
                 }
+
                 if (isMapped != PropertyDescriptor.Undefined)
                 {
                     if (desc.IsAccessorDescriptor())