… Jint/Native/Function/FunctionConstructor.cs
… Jint/Native/Function/FunctionPrototype.cs
… Jint/Native/Function/ScriptFunctionInstance.cs
… Jint/Native/Json/JsonSerializer.cs
… Jint/Native/Number/NumberConstructor.cs
… Jint/Native/Object/ObjectConstructor.cs
… Jint/Native/Object/ObjectInstance.cs
… Jint/Native/RegExp/RegExpConstructor.cs
… Jint/Native/RegExp/RegExpPrototype.cs
… Jint/Native/String/StringConstructor.cs
… Jint/Native/String/StringInstance.cs
… Jint/Native/String/StringPrototype.cs
… Jint/Native/Symbol/SymbolConstructor.cs
… Jint/Native/Symbol/SymbolPrototype.cs
… Jint/Runtime/Descriptors/IPropertyDescriptor .cs
… Jint/Runtime/Descriptors/PropertyDescriptor.cs
… Jint/Runtime/Descriptors/PropertyDescriptorExtensions.cs
… Jint/Runtime/Descriptors/Specialized/AllForbiddenPropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/ClrAccessDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/ConfigurableEnumerableWritablePropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/EnumerablePropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/NonConfigurablePropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/NonEnumerablePropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/NullConfigurationPropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/WritablePropertyDescriptor.cs
… Jint/Runtime/Environments/ObjectEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/Interop/ClrFunctionInstance.cs
… Jint/Runtime/Interop/NamespaceReference.cs
… Jint/Runtime/Interop/ObjectWrapper.cs