@lahma 2018-01-08
Thanks. End user API changes with IPropertyDescriptor being returned/taken as parameter (PropertyDescriptor works as before for defining things). Also the methods IsDataDescriptor() etc moved to be extension methods to help interface-only implementations. 

Feel free to merge if/when you find suitable, all tweaks naturally welcome.