@sebastienros 2018-01-08
I was just wondering if it would be better to introduce `[Flags]`. I know I have tried by the past, but I think it was more pain than this. Unless there are helper methods.