@lahma 2018-01-08
Yes, that's an alternative. Without interface it would marginally faster. Of course the flags need to be stored as field but getters could do bit checks. The extra Get and Set would be hard to model without properties though.

But I'm in no hurry with this one. The last merge brought the speed improvements already, this is tweaking and balancing now.